extend MainActivity and RecordingActivity from BaseActivity
