+++ recorder/src/main/java/com/mapbox/vision/recorder/RecordingActivity.kt
@@ -12,6 +12,9 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
 import com.mapbox.vision.VisionManager
+import com.mapbox.vision.common.BaseActivity
+import com.mapbox.vision.common.view.hide
+import com.mapbox.vision.common.view.show
 import com.mapbox.vision.mobile.core.interfaces.VisionEventsListener
 import com.mapbox.vision.mobile.core.models.Camera
 import com.mapbox.vision.mobile.core.models.Country
@@ -25,11 +28,11 @@ import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
 
-class RecordingActivity : AppCompatActivity() {
+class RecordingActivity : BaseActivity() {
 
     companion object {
-        private const val PERMISSIONS_REQUEST_CODE = 1
-        private const val PERMISSION_FOREGROUND_SERVICE = "android.permission.FOREGROUND_SERVICE"
+        private val BASE_SESSION_PATH = "${Environment.getExternalStorageDirectory().absolutePath}/Telemetry"
+        private val dateFormat = SimpleDateFormat("yyyy-MM-dd_HH-mm-ssZ", Locale.US)
     }
 
     private var isPermissionsGranted = false
@@ -60,33 +63,11 @@ class RecordingActivity : AppCompatActivity() {
         }
     }
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        super.onCreate(savedInstanceState)
+    override fun initViews() {
         setContentView(R.layout.activity_recording)
-
-        val board = SystemInfoUtils.getSnpeSupportedBoard()
-
-        if (!SupportedSnapdragonBoards.isBoardSupported(board)) {
-            val text =
-                Html.fromHtml("The device is not supported, you need <b>Snapdragon-powered</b> device with <b>OpenCL</b> support, more details at <b>https://www.mapbox.com/android-docs/vision/overview/</b>")
-            Toast.makeText(this, text, Toast.LENGTH_LONG).show()
-            VisionLogger.e(
-                "NotSupportedBoard",
-                "Current board is {\"$board\"}, Supported Boards: [${enumValues<SupportedSnapdragonBoards>().joinToString { it.name }}]; System Info: [${SystemInfoUtils.obtainSystemInfo()}]"
-            )
-            finish()
-            return
-        }
-
-        if (!allPermissionsGranted() && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            requestPermissions(getRequiredPermissions(), PERMISSIONS_REQUEST_CODE)
-        } else {
-            onPermissionsGranted()
-        }
     }
 
-    private fun onPermissionsGranted() {
+    public override fun onPermissionsGranted() {
         isPermissionsGranted = true
 
         root.setOnLongClickListener {
@@ -103,20 +84,6 @@ class RecordingActivity : AppCompatActivity() {
 
         tryToInitVisionManager()
     }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (allPermissionsGranted() && requestCode == PERMISSIONS_REQUEST_CODE) {
-            onPermissionsGranted()
-        }
-    }
-
-    private val BASE_SESSION_PATH = "${Environment.getExternalStorageDirectory().absolutePath}/Telemetry"
-    private val dateFormat = SimpleDateFormat("yyyy-MM-dd_HH-mm-ssZ", Locale.US)
 
     private fun tryToInitVisionManager() {
         if (isPermissionsGranted && !visionManagerWasInit) {
@@ -137,19 +104,6 @@ class RecordingActivity : AppCompatActivity() {
         tryToInitVisionManager()
     }
 
-    private fun toggleRecording() {
-        when (recording_view.state) {
-            RecordingView.State.NotRecording -> {
-                VisionManager.startRecording("$BASE_SESSION_PATH/${dateFormat.format(Date(System.currentTimeMillis()))}")
-                recording_view.state = RecordingView.State.Recording
-            }
-            RecordingView.State.Recording -> {
-                VisionManager.stopRecording()
-                recording_view.state = RecordingView.State.NotRecording
-            }
-        }
-    }
-
     override fun onStop() {
         super.onStop()
 
@@ -162,40 +116,16 @@ class RecordingActivity : AppCompatActivity() {
         }
     }
 
-    private fun allPermissionsGranted(): Boolean {
-        for (permission in getRequiredPermissions()) {
-            if (ContextCompat.checkSelfPermission(this, permission) != PackageManager.PERMISSION_GRANTED) {
-                // PERMISSION_FOREGROUND_SERVICE was added for targetSdkVersion >= 28, it is normal and always granted, but should be added to the Manifest file
-                // on devices with Android < P(9) checkSelfPermission(PERMISSION_FOREGROUND_SERVICE) can return PERMISSION_DENIED, but in fact it is GRANTED, so skip it
-                // https://developer.android.com/guide/components/services#Foreground
-                if (permission == PERMISSION_FOREGROUND_SERVICE) {
-                    continue
-                }
-                return false
+    private fun toggleRecording() {
+        when (recording_view.state) {
+            RecordingView.State.NotRecording -> {
+                VisionManager.startRecording("$BASE_SESSION_PATH/${dateFormat.format(Date(System.currentTimeMillis()))}")
+                recording_view.state = RecordingView.State.Recording
+            }
+            RecordingView.State.Recording -> {
+                VisionManager.stopRecording()
+                recording_view.state = RecordingView.State.NotRecording
             }
         }
-        return true
-    }
-
-    private fun getRequiredPermissions(): Array<String> {
-        return try {
-            val info = packageManager?.getPackageInfo(packageName, PackageManager.GET_PERMISSIONS)
-            val ps = info!!.requestedPermissions
-            if (ps != null && ps.isNotEmpty()) {
-                ps
-            } else {
-                emptyArray()
-            }
-        } catch (e: Exception) {
-            emptyArray()
-        }
-    }
-
-    private fun View.hide() {
-        visibility = View.GONE
-    }
-
-    private fun View.show() {
-        visibility = View.VISIBLE
     }
 }
