--- recorder/src/main/res/drawable/ic_watermark_logo.xml
@@ -1,70 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="104dp"
-    android:height="24dp"
-    android:viewportWidth="104"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M76.579,15.071C74.896,15.071 73.55,13.724 73.55,11.956L73.55,11.872C73.55,10.103 74.896,8.756 76.579,8.756C78.262,8.756 79.609,10.188 79.609,11.956C79.609,13.724 78.262,15.071 76.579,15.071ZM77,6.651C75.738,6.651 74.476,7.157 73.55,8.167L73.55,3.199C73.55,3.031 73.382,2.863 73.213,2.863L71.446,2.863C71.278,2.863 71.11,3.031 71.11,3.199L71.11,16.671C71.11,16.839 71.278,17.008 71.446,17.008L73.213,17.008C73.382,17.008 73.55,16.839 73.55,16.671L73.55,15.745C74.476,16.671 75.654,17.176 77,17.26C79.693,17.26 81.965,14.903 81.965,11.956C81.965,9.009 79.693,6.651 77,6.651Z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-  <path
-      android:pathData="M88.36,15.071C86.677,15.071 85.331,13.64 85.331,11.872C85.331,10.103 86.677,8.672 88.36,8.672C90.043,8.672 91.389,10.103 91.389,11.872C91.389,13.724 90.043,15.071 88.36,15.071ZM88.36,6.651C85.415,6.651 82.975,9.009 82.975,11.956C82.975,14.903 85.415,17.26 88.36,17.26C91.305,17.26 93.746,14.903 93.746,11.956C93.746,9.009 91.305,6.651 88.36,6.651Z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-  <path
-      android:pathData="M50.662,15.071C48.979,15.071 47.632,13.64 47.632,11.872C47.632,10.103 48.979,8.672 50.662,8.672C52.345,8.672 53.691,10.019 53.691,11.787L53.691,11.872C53.607,13.724 52.26,15.071 50.662,15.071ZM55.711,6.988L53.943,6.988C53.775,6.988 53.607,7.157 53.607,7.325L53.607,8.251C52.681,7.325 51.503,6.82 50.157,6.736C47.464,6.736 45.192,9.093 45.192,12.04C45.192,14.987 47.38,17.344 50.157,17.344C51.419,17.344 52.681,16.839 53.607,15.829L53.607,16.755C53.607,16.923 53.775,17.092 53.943,17.092L55.711,17.092C55.879,17.092 56.047,16.923 56.047,16.755L56.047,7.325C55.963,7.072 55.879,6.988 55.711,6.988Z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-  <path
-      android:pathData="M64.041,15.071C62.358,15.071 61.012,13.724 61.012,11.956L61.012,11.872C61.012,10.103 62.358,8.756 64.041,8.756C65.724,8.756 67.071,10.188 67.071,11.956C67.071,13.724 65.724,15.071 64.041,15.071ZM64.462,6.651C63.2,6.651 61.938,7.157 61.012,8.167L61.012,7.241C61.012,7.072 60.844,6.904 60.675,6.904L58.908,6.904C58.74,6.904 58.572,7.072 58.572,7.241L58.572,20.712C58.572,20.881 58.74,21.049 58.908,21.049L60.675,21.049C60.844,21.049 61.012,20.881 61.012,20.712L61.012,15.745C61.938,16.671 63.116,17.176 64.462,17.26C67.155,17.26 69.427,14.903 69.427,11.956C69.427,9.009 67.155,6.651 64.462,6.651Z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-  <path
-      android:pathData="M40.059,6.651C38.713,6.651 37.535,7.409 37.03,8.588C36.356,7.409 35.094,6.651 33.748,6.651C32.654,6.651 31.644,7.157 31.055,7.999L31.055,7.157C31.055,6.988 30.887,6.82 30.719,6.82L28.951,6.82C28.783,6.82 28.615,6.988 28.615,7.157L28.615,16.587C28.615,16.755 28.783,16.923 28.951,16.923L30.719,16.923C30.887,16.923 31.055,16.755 31.055,16.587L31.055,10.945C31.139,9.683 31.981,8.672 33.075,8.672C34.169,8.672 35.094,9.598 35.094,10.861L35.094,16.587C35.094,16.755 35.263,16.923 35.431,16.923L37.198,16.923C37.366,16.923 37.535,16.755 37.535,16.587L37.535,10.693C37.703,9.514 38.544,8.672 39.554,8.672C40.648,8.672 41.574,9.598 41.574,10.861L41.574,16.587C41.574,16.755 41.742,16.923 41.91,16.923L43.677,16.923C43.846,16.923 44.014,16.755 44.014,16.587L44.014,10.103C43.677,8.251 42.079,6.651 40.059,6.651Z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-  <path
-      android:pathData="M103.17,16.502L100.141,11.872L103.17,7.325C103.254,7.157 103.254,6.988 103.086,6.988C103.002,6.988 103.002,6.904 102.918,6.904L100.898,6.904C100.73,6.904 100.562,6.988 100.562,7.157L98.71,10.103L96.943,7.157C96.859,6.988 96.691,6.904 96.607,6.904L94.503,6.904C94.335,6.904 94.251,7.072 94.251,7.157C94.251,7.241 94.251,7.241 94.335,7.325L97.364,11.872L94.335,16.502C94.251,16.671 94.251,16.839 94.419,16.839C94.503,16.839 94.503,16.923 94.587,16.923L96.607,16.923C96.775,16.923 96.943,16.839 96.943,16.671L98.71,13.64L100.478,16.671C100.562,16.839 100.73,16.923 100.814,16.923L102.834,16.923C103.002,16.923 103.086,16.755 103.086,16.671C103.17,16.587 103.17,16.587 103.17,16.502Z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-  <path
-      android:pathData="M17.086,14.818C13.131,18.776 5.979,17.513 5.979,17.513C5.979,17.513 4.717,10.44 8.672,6.399C10.859,4.21 14.562,4.294 16.834,6.567C19.106,8.841 19.274,12.629 17.086,14.818ZM11.617,0.337C5.222,0.337 0.004,5.557 0.004,11.956C0.004,18.355 5.222,23.575 11.617,23.575C18.012,23.575 23.229,18.355 23.229,11.956C23.229,5.557 18.012,0.337 11.617,0.337Z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-  <path
-      android:pathData="M12.879,7.157l-1.094,2.357l-2.356,1.095l2.356,1.179l1.094,2.357l1.178,-2.357l2.356,-1.179l-2.356,-1.095z"
-      android:strokeAlpha="0.4"
-      android:strokeWidth="1"
-      android:fillColor="#FFFFFF"
-      android:fillType="nonZero"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.4"/>
-</vector>
